initial commit
The strategy here is simple: copy the map iterator implementation
from runtime/hashmap.go, but seed it with a stronger random value.
Then simply return the key/value at the starting position of the
iterator.

Unfortunately, there seems to be a problem: Go's iterator seems
to heavily favor the first key/value in the map. Further work is
required to find out why. The only clue is that in maps with fewer
than 8 elements, all but the first are selected 12.5% of the time;
i.e. for a 5-element map, the first element is selected 50% of the
time, and for a 3-element map, it is selected 75% of the time.