achieve uniformity
All functions now select map keys/values with proper uniformity.
The favoring was due to Go's map iteration algorithm: it starts
at a random index in the map's buckets, and then moves forward
linearly until it finds a non-empty bucket. As a result, if a
map contains only two elements, the first is far more likely to
be selected than the second, since most of the initial random
indices will land on an empty bucket and progress until they hit
the first element. The second element, however, can *only* be
selected if it the initial random index lands on it exactly.

This algorithm has been converted to the following: instead of
progressing linearly after landing on a empty bucket, we simply
retry with a new random index. The runtime is therefore O(n/k),
where n is the number of buckets and k is the number of elements.
Actually, the worst case is unbounded, since no effort is made
to avoid picking the same index twice. This could be improved by
using a random permutation of possible indices. It is worth noting,
however, that maps are almost never <50% full; as such, this
algorithm rarely needs to restart more than once.

Now that it works, the next step is to make it fast. FastKey/Val
are already acceptably fast compared to the builtin map. The
cryptographically-strong Key/Val are quite a bit slower, mostly
due to the overhead of reading /dev/urandom. There are a few ways
to mitigate this.

Finally, I would like to implement a function that iterates over a
map in random order, a la rand.Perm. Hopefully this won't require
too much additional work. I'm thinking the syntax will look like:

    randmap.Iter(m, func(key, val int) {
        // use key and val
    })

... which will of course require some more reflection magic.