fix overflow bug
The previous implementation suffered from a bug which caused it to
never select elements from overflow buckets. Note that since overflow
buckets are only created in the event of a hash collision, this bug
is nondeterministic.

Unfortunately, the requirement that we check overflow buckets makes
it much harder to achieve good performance. The problem is that we
don't know which buckets have overflows in advance, meaning we can't
select uniformly among them.

There are two ways to solve this: we can iterate through all the
buckets+overflows and use Fisher-Yates to select one at random, or
we can expand our "random dart throw" into the 2nd dimension, i.e.
a rectangle with one side being the number of buckets and the other
being the maximum number of overflows. I have taken the latter
approach, which means the runtime is now O(L * (1+k/n)), where L is
the maximum number of overflows. (The +1 comes from the fact that
we must iterate L times to access the overflow bucket.)

Lastly: the overflow bucket is selected via modulus. Some people
have complained that modulo isn't random enough, but it's sufficient
for our purposes. It's easy to change later if it's a real concern.