implement (Fast)Iter
This implementation is unfortunately quite slow. The primary causes
are the use of reflection and the allocation of a slice in randPerm.
The slice allocation can possibly be removed via a pseudorandom
permutation generator, potentially using a block cipher. As for the
reflection... some more unsafe magic may be required, and even then
you'll be missing out on lots of potential for compiler optimization.
So it'll be a tradeoff; vs. the flattening approach, you trade some
speed for O(1) space.

I also noticed that I can simplify the implementations of Key and
Val by using mapaccessi in place of randIter. That will come next.