make randIter constant space
randIter now uses a Feistel network to generate permutations in
constant space. blake2b is used as the round function, which is
probably an awful choice -- randIter is still quite slow. The
generator permutes up to the nearest power of 4, and we have to
throw out anything too large, so that adds a worst-case factor of
4 to our already-miss-prone map access algorithm.

Also, the generator is seeded with a uint32, which is probably not
enough entropy. Easy to make it a [16]byte or something later.
One known problem with these generators is that they cannot hope
to enumerate the entire permutation space, since their seed space
is typically much smaller than factorial(n).

This took a lot of trial and error to get working! I had a lot
of trouble with 0 always permuting to 0. Finally I realized why:
I was using the same subkey for every round, and thus the same
pseudorandom function (Feistel round functions are supposed to be
independent). I addressed this by incrementing the seed for each
round, and suddenly the distribution became uniform.

I would really appreciate some more thorough analysis of this code
(preferably by a professional), but it seems correct enough for me
to start optimizing it. Ideally, I could use random permutations
to achieve an upper bound on Key and Val:

    g := newGenerator(space, seed)
    g.Iter(func(r uint32) {
        if mapaccessi(t, h, etc...) {
            break // out of Iter, somehow
        }
    })
    return key

This is pretty funny when you consider it's basically identical to
the stdlib-init approach, except properly random!