make Iter a proper iterator
This allows you to break out of the iteration, whereas the old
"functional" style did not. Perhaps I could have added a return
value to the supplied function, but that felt ugly.

I said in a previous commit that the generator could be used for
Key and Val, but this is not feasible because the generator is so
slow. Kind of a shame, but oh well.

I'm pretty irked that there is no way to simply overwrite the value
pointer with the one we get via map iteration. The standard library
accomplishes this using a private function called typedmemmove, but
it's not available to us. So we have to use it indirectly via the
reflect package. I doubt this hurts performance much, though; Iter
is already so slow that the reflect overhead is negligible.